Update to client version 63
This is an attempt to manually merge all the functional changes from
Soren Wilkening's pull request. As I mentioned in the issue, I had to
figure out a way to effectively ignore all the formatting changes in
order to find and review the changes in functionality.

I happened to notice that it looked like Soren used package.skeleton()
to write all his changes to files. I did the same with the existing
version of the package, and that essentially formatted the current
package code the same as Soren's.

The differences are mainly additions to incoming message types,
eWrappers, and order object components. Soren removed some
functionality but it was not clear he did, so I did not remove it.

I did remove all the calls to .Internal() that Soren added, because
the package will not pass R CMD check with them.

Closes #9.

Adds new incoming message types and eWrappers:

  MARKET_DATA_TYPE = "58"
  COMMISSION_REPORT = "59"
  POSITION_DATA = "61"
  POSITION_END = "62"
  ACCOUNT_SUMMARY = "63"
  ACCOUNT_SUMMARY_END = "64"
  VERIFY_MESSAGE_API = "65"
  VERIFY_COMPLETED = "66"
  DISPLAY_GROUP_LIST = "67"
  DISPLAY_GROUP_UPDATED = "68"

Adds new order components:

  exemptCode
  trailingPercent
  scaleTable
  activeStartTime
  activeStopTime
  hedgeType
  hedgeParam
  optOutSmartRouting
  algoStrategy
  conId
  tradingClass

  deltaNeutralConId
  deltaNeutralSettlingFirm
  deltaNeutralClearingAccount
  deltaNeutralClearingIntent
  deltaNeutralOpenClose
  deltaNeutralShortSale
  deltaNeutralShortSaleSlot
  deltaNeutralDesignatedLocation

  scalePriceAdjustValue
  scalePriceAdjustInterval
  scaleProfitOffset
  scaleAutoReset
  scaleInitPosition
  scaleInitFillQty
  scaleRandomPercent

  smartComboRoutingParams
  smartComboRoutingParamsCount
  orderComboLegs
  orderComboLegsCount
  comboLegsCount
  orderMiscOptions