Correct checks for twsPlayback connections
Calling twsConnect() on a file did not work because the twsConnect()
call returns a 'twsplay' classed object and twsCALLBACK() only checked
for 'twsPlayback' classed objects.

Thanks to Laser Yuan for the report and patch.

Make the same change to reqMktDepth(), though it worked because it
only checked for 'twsPlayback' classed objects.

Fixes #6.