Add startApi() to create the API connection
The handshake between the server and client changed somewhere between
version 47 and 63. Now it needs a START_API message after the client
and server versions have been established.

The server responds to the START_API message with
  1. a managed account message,
  2. a next valid ID, and
  3. messages with the status of the data farms

Soren did not store the server response, so this commit does not
recreate his PR verbatim. One major difference: this change does not
remove the ewrapper, so the next valid ID is still stored in the
ewrapper environment instead of the global environment (as it does in
Soren's PR).

See #19, see #9.