Add reqMktDataType
Allow users to request live, frozen, delayed, or delayed frozen market
data from TWS before calling reqMktData().

The API documentation says a successful switch to non-live data should
be indicated by a callback to eWrapper$marketDataType() with the ticker
ID of the market data request which is returning a different type of
data. This commit does not provide that functionality.

Fixes #18.